Projekt löschen. Neues Projekt ablegen. Projekt Info abrufen.

+++ src/main/java/de/szut/lf8_project/coworker/CoworkerEntity.java
@@ -12,7 +12,7 @@ import java.util.*;
 public class CoworkerEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int coworkerId;
+    private long coworkerId;
     private String name;
     private String surname;
     private long age;
