+++ src/main/java/de/szut/lf8_project/coworker/dto/CoworkerGetDto.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 
 @Data
 public class CoworkerGetDto {
-    private int coworkerId;
+    private long coworkerId;
     private String name;
     private String surname;
     private long age;
